Always compact ETCD before evaluating reclaimable space
The problem was that we did not compact ETCD before evaluating how much
space can bee reclaimed. Most of the time, the reclaimable space is not
beyond threshold leading to ETCD going above 80% use and triggering ETCD
shield to deny PipelineRun.

Always perform compaction, which does not have a penalty, i.e. does not
cause ETCD member from being unavailable for a few minutes like defrag.
Pass the --physical option to wait for compaction to finish before
moving to evaluation of reclaimable space. Then evaluate base if defrag
is needed, as it was done before.

Also simplify the defrag conditions, no need to have nested if, we want
to defrag if we are above 15% fragmentation or if 1GB can be reclaimed.

KFLUXINFRA-1644

Signed-off-by: Hugo Arès <hares@redhat.com>